refactoring location of current deposition and particle classes.
