+++ PyPIC3D/boris.py
@@ -2,7 +2,7 @@ import jax
 from jax import jit
 import jax.numpy as jnp
 
-from PyPIC3D.shapes import get_first_order_weights, get_second_order_weights
+from PyPIC3D.deposition.shapes import get_first_order_weights, get_second_order_weights
 from PyPIC3D.utils import wrap_around
 
 @jit
