--- PyPIC3D/particle.py
@@ -1,747 +1,0 @@
-import numpy as np
-import jax
-from jax import jit
-from functools import partial
-import jax.numpy as jnp
-from jax.tree_util import register_pytree_node_class
-
-from PyPIC3D.utils import vth_to_T, plasma_frequency, debye_length, T_to_vth
-
-def grab_particle_keys(config):
-    """
-    Extracts and returns a list of keys from the given configuration dictionary
-    that start with the prefix 'particle'.
-
-    Args:
-        config (dict): A dictionary containing configuration keys and values.
-
-    Returns:
-        list: A list of keys from the configuration dictionary that start with 'particle'.
-    """
-    particle_keys = []
-    for key in config.keys():
-        if key[:8] == 'particle':
-            particle_keys.append(key)
-    return particle_keys
-
-def load_particles_from_toml(config, simulation_parameters, world, constants):
-    """
-    Load particle data from a TOML file and initialize particle species.
-    Args:
-        config (dict): Dictionary containing configuration keys and values.
-        simulation_parameters (dict): Dictionary containing simulation parameters.
-        world (dict): Dictionary containing world parameters such as 'x_wind', 'y_wind', 'z_wind', 'dx', 'dy', 'dz'.
-        constants (dict): Dictionary containing constants such as 'kb'.
-    Returns:
-        list: A list of particle_species objects initialized with the data from the TOML file.
-
-    The function reads particle configuration from the provided TOML file, initializes particle properties such as
-    position, velocity, charge, mass, and temperature. It also handles loading initial positions and velocities from
-    external sources if specified in the TOML file. The particles are then appended to a list and returned.
-    """
-
-    x_wind = world['x_wind']
-    y_wind = world['y_wind']
-    z_wind = world['z_wind']
-    # get the world dimensions
-    Nx = world['Nx']
-    Ny = world['Ny']
-    Nz = world['Nz']
-    # get the number of grid points in each dimension
-    dx = world['dx']
-    dy = world['dy']
-    dz = world['dz']
-    dt = world['dt']
-    # get spatial and temporal resolution
-    kb = constants['kb']
-    eps = constants['eps']
-    C   = constants['C']
-    # get the constants
-
-    i = 0
-    # initialize the random number generator key
-    # this is used to generate random numbers for the initial positions and velocities of the particles
-    # it is incremented by 3 for each particle species to ensure different random numbers for each species
-    particles = []
-    particle_keys = grab_particle_keys(config)
-    # get the particle keys from the config dictionary
-
-    weight = compute_macroparticle_weight(config, particle_keys, simulation_parameters, world, constants)
-    # scale the particle weight by the debye length to prevent numerical heating
-    # this is done by computing the total debye length of the plasma and scaling the particle weight accordingly
-
-
-    for toml_key in particle_keys:
-        key1, key2, key3 = jax.random.key(i), jax.random.key(i+1), jax.random.key(i+2)
-        i += 3
-        # build the particle random number generator keys
-        particle_name = config[toml_key]['name']
-        print(f"\nInitializing particle species: {particle_name}")
-        charge=config[toml_key]['charge']
-        mass=config[toml_key]['mass']
-
-        if 'N_particles' in config[toml_key]:
-            N_particles=config[toml_key]['N_particles']
-            N_per_cell = N_particles / (world['Nx'] * world['Ny'] * world['Nz'])
-        elif "N_per_cell" in config[toml_key]:
-            N_per_cell = config[toml_key]["N_per_cell"]
-            N_particles = int(N_per_cell * world['Nx'] * world['Ny'] * world['Nz'])
-        # set the number of particles in the species
-
-        if 'temperature' in config[toml_key]:
-            T=config[toml_key]['temperature']
-            vth = T_to_vth(T, mass, kb)
-        elif 'vth' in config[toml_key]:
-            vth = config[toml_key]['vth']
-            T = vth_to_T(vth, mass, kb)
-        else:
-            T = 1.0
-            vth = T_to_vth(T, mass, kb)
-        # set the temperature of the particle species
-
-        Tx = read_value('Tx', toml_key, config, T)
-        Ty = read_value('Ty', toml_key, config, T)
-        Tz = read_value('Tz', toml_key, config, T)
-        # set the temperature of the particle species in each dimension
-
-        xmin = read_value('xmin', toml_key, config, -x_wind / 2)
-        xmax = read_value('xmax', toml_key, config, x_wind / 2)
-        ymin = read_value('ymin', toml_key, config, -y_wind / 2)
-        ymax = read_value('ymax', toml_key, config, y_wind / 2)
-        zmin = read_value('zmin', toml_key, config, -z_wind / 2)
-        zmax = read_value('zmax', toml_key, config, z_wind / 2)
-        # set the bounds for the particle species
-        x, y, z, vx, vy, vz = initial_particles(N_per_cell, N_particles, xmin, xmax, ymin, ymax, zmin, zmax, mass, Tx, Ty, Tz, kb, key1, key2, key3)
-        # initialize the positions and velocities of the particles
-
-        x_bc = 'periodic'
-        if 'x_bc' in config[toml_key]:
-            assert config[toml_key]['x_bc'] in ['periodic', 'reflecting'], f"Invalid x boundary condition: {config[toml_key]['x_bc']}"
-            x_bc = config[toml_key]['x_bc']
-        y_bc = 'periodic'
-        if 'y_bc' in config[toml_key]:
-            assert config[toml_key]['y_bc'] in ['periodic', 'reflecting'], f"Invalid y boundary condition: {config[toml_key]['y_bc']}"
-            y_bc = config[toml_key]['y_bc']
-        z_bc = 'periodic'
-        if 'z_bc' in config[toml_key]:
-            assert config[toml_key]['z_bc'] in ['periodic', 'reflecting'], f"Invalid z boundary condition: {config[toml_key]['z_bc']}"
-            z_bc = config[toml_key]['z_bc']
-        # set the boundary conditions for the particle species
-
-        x = load_initial_positions('initial_x', config, toml_key, x, N_particles, dx, Nx, key1)
-        y = load_initial_positions('initial_y', config, toml_key, y, N_particles, dy, Ny, key2)
-        z = load_initial_positions('initial_z', config, toml_key, z, N_particles, dz, Nz, key3)
-        # load the initial positions of the particles from the toml file, if specified
-        # otherwise, use the initialized positions
-        vx = load_initial_velocities('initial_vx', config, toml_key, vx, N_particles)
-        vy = load_initial_velocities('initial_vy', config, toml_key, vy, N_particles)
-        vz = load_initial_velocities('initial_vz', config, toml_key, vz, N_particles)
-        # load the initial velocities of the particles from the toml file, if specified
-        # otherwise, use the initialized velocities
-
-        # Calculate the temperature from the velocities if not explicitly set
-        if 'temperature' not in config[toml_key]:
-            T = (mass / (3 * kb * N_particles)) * (
-                jnp.sum(vx ** 2) + jnp.sum(vy ** 2) + jnp.sum(vz ** 2)
-            )
-
-        if "weight" in config[toml_key]:
-            weight = config[toml_key]['weight']
-            # set the weight of the particles, if specified in the toml file
-        elif 'ds_per_debye' in config[toml_key]:  # account for anisotropic grid spacings via ds2
-            ds_per_debye = config[toml_key]['ds_per_debye']
-            
-            ds2 = 0
-            for d in [dx, dy, dz]:
-                if d != 1:
-                    ds2 += d**2
-
-            if ds2 == 0:
-                raise ValueError(
-                    "Invalid configuration for 'ds_per_debye': at least one of dx, dy, dz must differ from 1."
-                )
-            weight = (x_wind*y_wind*z_wind * eps * kb * T)  / (N_particles * charge**2 * ds_per_debye**2 * ds2)
-            # weight the particles by the debye length and the number of particles
-
-        update_pos = read_value('update_pos', toml_key, config, True)
-        update_v = read_value('update_v', toml_key, config, True)
-        update_vx = read_value('update_vx', toml_key, config, True)
-        update_vy = read_value('update_vy', toml_key, config, True)
-        update_vz = read_value('update_vz', toml_key, config, True)
-        update_x = read_value('update_x', toml_key, config, True)
-        update_y = read_value('update_y', toml_key, config, True)
-        update_z = read_value('update_z', toml_key, config, True)
-
-        particle = particle_species(
-            name=particle_name,
-            N_particles=N_particles,
-            charge=charge,
-            mass=mass,
-            T=T,
-            x1=x,
-            x2=y,
-            x3=z,
-            v1=vx,
-            v2=vy,
-            v3=vz,
-            xwind=x_wind,
-            ywind=y_wind,
-            zwind=z_wind,
-            dx=dx,
-            dy=dy,
-            dz=dz,
-            weight=weight,
-            x_bc=x_bc,
-            y_bc=y_bc,
-            z_bc=z_bc,
-            update_vx=update_vx,
-            update_vy=update_vy,
-            update_vz=update_vz,
-            update_x=update_x,
-            update_y=update_y,
-            update_z=update_z,
-            update_pos=update_pos,
-            update_v=update_v,
-            shape=simulation_parameters['shape_factor'],
-            dt=dt
-        )
-        particles.append(particle)
-
-        pf = plasma_frequency(particle, world, constants)
-        dl = debye_length(particle, world, constants)
-        print(f"Number of particles: {N_particles}")
-        print(f"Number of particles per cell: {N_per_cell}")
-        print(f"x, y, z boundary conditions: {x_bc}, {y_bc}, {z_bc}")
-        print(f"Charge: {charge}")
-        print(f"Mass: {mass}")
-        print(f"Temperature: {T}")
-        print(f"Thermal Velocity: {vth}")
-        print(f"Particle Kinetic Energy: {particle.kinetic_energy()}")
-        print(f"Particle Species Plasma Frequency: {pf}")
-        print(f"Time Steps Per Plasma Period: {(1 / (dt * pf) )}")
-        print(f"Particle Species Debye Length: {dl}")
-        print(f"Particle Weight: {weight}")
-        print(f"Particle Species Scaled Charge: {particle.get_charge()}")
-        print(f"Particle Species Scaled Mass: {particle.get_mass()}")
-
-    return particles
-
-
-def read_value(param, key, config, default_value):
-    """
-    Reads a value from a nested dictionary structure and returns it if it exists;
-    otherwise, returns a default value.
-
-    Args:
-        param (str): The parameter name to look for in the nested dictionary.
-        key (str): The key in the outer dictionary where the nested dictionary is located.
-        config (dict): The configuration dictionary containing nested dictionaries.
-        default_value (Any): The value to return if the parameter is not found.
-
-    Returns:
-        Any: The value associated with `param` in `config[key]` if it exists,
-             otherwise `default_value`.
-    """
-    if param in config[key]:
-        print(f'Reading user defined {param}')
-        return config[key][param]
-    else:
-        return default_value
-
-
-def load_initial_positions(param, config, key, default, N_particles, ds, ns, key1):
-    """
-    Load initial positions for particles based on the provided configuration.
-
-    This function checks if a specific parameter exists in the configuration
-    under the given key. If the parameter exists and is a string, it loads
-    the data from an external source. If the parameter exists and is a number,
-    it creates an array filled with that value. If the parameter does not
-    exist, it returns the default value.
-
-    Args:
-        param (str): The name of the parameter to look for in the configuration.
-        config (dict): The configuration dictionary containing parameters and values.
-        key (str): The key in the configuration dictionary under which the parameter is stored.
-        default (Any): The default value to return if the parameter is not found.
-        N_particles (int): The number of particles, used to determine the size of the array.
-        ds (float): The spatial resolution, used to add noise to the particle positions.
-        key1 (jax.random.PRNGKey): The random key for generating random numbers.
-
-    Returns:
-        jax.numpy.ndarray or Any: An array of particle positions if the parameter is found,
-        or the default value if the parameter is not found.
-    """
-    if param in config[key]:
-        if isinstance(config[key][param], str):
-            print(f"Loading {param} from external source: {config[key][param]}")
-            return jnp.load(config[key][param])
-            # if the value is a string, load it from an external source
-        else:
-            #return jnp.full(N_particles, config[key][param])
-            val = config[key][param]
-            if ns == 1:
-                return val * jnp.ones(N_particles)
-            else:
-                return jax.random.uniform(key1, shape=(N_particles,), minval=val-(ds/2), maxval=val+(ds/2))
-            # if the value is a number, fill the array with that value with some noise in the subcell position
-    else:
-        return default
-        # return the default value if the parameter is not found
-
-def load_initial_velocities(param, config, key, default, N_particles):
-    """
-    Load initial velocities for particles based on the provided configuration.
-
-    This function checks if a specific parameter exists in the configuration
-    dictionary under the given key. Depending on the type of the parameter's
-    value, it either loads data from an external source or initializes an array
-    with a specified value. If the parameter is not found, a default value is returned.
-
-    Args:
-        param (str): The name of the parameter to look for in the configuration.
-        config (dict): A dictionary containing configuration data.
-        key (str): The key in the configuration dictionary where the parameter is located.
-        default (float or jnp.ndarray): The default value to return if the parameter is not found.
-        N_particles (int): The number of particles, used to determine the size of the array.
-
-    Returns:
-        jnp.ndarray: An array of initial velocities for the particles. If the parameter
-        is a string, the array is loaded from an external source. If the parameter is
-        a number, the array is filled with that value plus the default. If the parameter
-        is not found, the default value is returned.
-    """
-    if param in config[key]:
-        if isinstance(config[key][param], str):
-            print(f"Loading {param} from external source: {config[key][param]}")
-            return jnp.load(config[key][param])
-            # if the value is a string, load it from an external source
-        else:
-            return jnp.full(N_particles, config[key][param]) + default
-            # if the value is a number, fill the array with that value
-    else:
-        return default
-        # return the default value if the parameter is not found
-
-def compute_macroparticle_weight(config, particle_keys, simulation_parameters, world, constants):
-
-    x_wind = world['x_wind']
-    y_wind = world['y_wind']
-    z_wind = world['z_wind']
-    # get the world dimensions
-    dx = world['dx']
-    dy = world['dy']
-    dz = world['dz']
-    # get the world resolution
-    kb = constants['kb']
-    eps = constants['eps']
-    # get the constants
-
-    if simulation_parameters['ds_per_debye']: # scale the particle weight by the debye length to prevent numerical heating
-        ds_per_debye = simulation_parameters['ds_per_debye']
-        # get the number of grid points per debye length
-        inverse_total_debye = 0
-
-        for toml_key in particle_keys:
-            N_particles = config[toml_key]['N_particles']
-            charge = config[toml_key]['charge']
-            mass = config[toml_key]['mass']
-            # get the charge and mass of the particle species
-            if 'temperature' in config[toml_key]:
-                T=config[toml_key]['temperature']
-            elif 'vth' in config[toml_key]:
-                T = vth_to_T(config[toml_key]['vth'], mass, kb)
-            # get the temperature of the particle species
-
-            inverse_total_debye += jnp.sqrt( N_particles / (x_wind * y_wind * z_wind) / (eps * kb * T) ) * jnp.abs(charge)
-            # get the inverse debye length before macroparticle weighting
-
-        ds2 = 0
-        for d in [dx, dy, dz]:
-            if d != 1:
-                ds2 += d**2
-            
-
-        weight = 1 / (ds2) / (ds_per_debye**2) / inverse_total_debye
-        # weight the particles by the total debye length of the plasma
-
-    else:
-        weight = 1.0 # default to single particle weight
-
-    return weight
-
-def initial_particles(N_per_cell, N_particles, minx, maxx, miny, maxy, minz, maxz, mass, Tx, Ty, Tz, kb, key1, key2, key3):
-    """
-    Initializes the velocities and positions of the particles.
-
-    Args:
-        N_particles (int): The number of particles.
-        minx (float): The minimum value for the x-coordinate of the particles' positions.
-        maxx (float): The maximum value for the x-coordinate of the particles' positions.
-        miny (float): The minimum value for the y-coordinate of the particles' positions.
-        maxy (float): The maximum value for the y-coordinate of the particles' positions.
-        minz (float): The minimum value for the z-coordinate of the particles' positions.
-        maxz (float): The maximum value for the z-coordinate of the particles' positions.
-        mass (float): The mass of the particles.
-        T (float): The temperature of the system.
-        kb (float): The Boltzmann constant.
-        key (jax.random.PRNGKey): The random key for generating random numbers.
-
-    Returns:
-        x (jax.numpy.ndarray): The x-coordinates of the particles' positions.
-        y (jax.numpy.ndarray): The y-coordinates of the particles' positions.
-        z (jax.numpy.ndarray): The z-coordinates of the particles' positions.
-        v_x (numpy.ndarray): The x-component of the particles' velocities.
-        v_y (numpy.ndarray): The y-component of the particles' velocities.
-        v_z (numpy.ndarray): The z-component of the particles' velocities.
-    """
-
-    # if N_per_cell < 1:
-    x = jax.random.uniform(key1, shape = (N_particles,), minval=minx, maxval=maxx)
-    y = jax.random.uniform(key2, shape = (N_particles,), minval=miny, maxval=maxy)
-    z = jax.random.uniform(key3, shape = (N_particles,), minval=minz, maxval=maxz)
-        # initialize the positions of the particles
-    # else:
-        # x = jnp.repeat(jax.random.uniform(key1, shape=(N_particles // N_per_cell,), minval=minx, maxval=maxx), N_per_cell)
-        # y = jnp.repeat(jax.random.uniform(key2, shape=(N_particles // N_per_cell,), minval=miny, maxval=maxy), N_per_cell)
-        # z = jnp.repeat(jax.random.uniform(key3, shape=(N_particles // N_per_cell,), minval=minz, maxval=maxz), N_per_cell)
-        # initialize the positions of the particles, giving every N_per_cell particles the same position
-    #std = jnp.sqrt( kb * T / mass )
-    std_x = T_to_vth( Tx, mass, kb )
-    std_y = T_to_vth( Ty, mass, kb )
-    std_z = T_to_vth( Tz, mass, kb )
-    v_x = np.random.normal(0, std_x, N_particles)
-    v_y = np.random.normal(0, std_y, N_particles)
-    v_z = np.random.normal(0, std_z, N_particles)
-    # initialize the particles with a maxwell boltzmann distribution.
-    return x, y, z, v_x, v_y, v_z
-
-@jit
-def compute_index(x, dx, window):
-    """
-    Compute the index of a position in a discretized space.
-
-    Args:
-        x (float or ndarray): The position(s) to compute the index for.
-        dx (float): The discretization step size.
-
-    Returns:
-        int or ndarray: The computed index/indices as integer(s).
-    """
-    scaled_x = x + window/2
-    return jnp.floor( scaled_x / dx).astype(int)
-
-
-@partial(jit, static_argnames=("periodic", "reflecting"))
-def apply_axis_boundary_condition(x, v, wind, half_wind, periodic, reflecting):
-    """
-    Apply boundary conditions to particle positions and velocities along a single axis.
-    This function handles three types of boundary conditions: periodic, reflecting, and open.
-    Particles that cross the boundary are treated according to the specified condition type.
-    Args:
-        x (jnp.ndarray): Particle positions along the axis.
-        v (jnp.ndarray): Particle velocities along the axis.
-        wind (float): The width of the domain (full extent from -half_wind to half_wind).
-        half_wind (float): Half the domain width, used to determine boundary positions.
-        periodic (bool): If True, apply periodic boundary conditions (particles wrap around).
-        reflecting (bool): If True, apply reflecting boundary conditions (particle velocities reverse).
-                          Only evaluated if periodic is False.
-    Returns:
-        tuple: A tuple of (x_out, v_out) where:
-            - x_out (jnp.ndarray): Updated particle positions after applying boundary conditions.
-            - v_out (jnp.ndarray): Updated particle velocities after applying boundary conditions.
-    Notes:
-        - Periodic BC: Particles that exceed ±half_wind are wrapped to the opposite side.
-        - Reflecting BC: Particles at boundaries have their velocities reversed (x position unchanged).
-        - Open BC (default): Particles and velocities pass through unchanged.
-    """
-
-    def periodic_bc(state):
-        x_in, v_in = state
-        x_out = x_in + wind * (x_in < -half_wind) - wind * (x_in > half_wind)
-        return x_out, v_in
-    # if periodic is True, apply periodic boundary conditions by wrapping positions around the domain
-
-    def reflecting_bc(state):
-        x_in, v_in = state
-        v_out = jnp.where((x_in >= half_wind) | (x_in <= -half_wind), -v_in, v_in)
-        return x_in, v_out
-    # if reflecting is True, apply reflecting boundary conditions by reversing velocities at the boundaries
-
-    def identity_bc(state):
-        return state
-    # if neither periodic nor reflecting, return the positions and velocities unchanged
-
-    return jax.lax.cond(
-        periodic,
-        periodic_bc,
-        lambda state: jax.lax.cond(reflecting, reflecting_bc, identity_bc, state),
-        (x, v),
-    )
-
-@register_pytree_node_class
-class particle_species:
-    """
-    Class representing a species of particles in a simulation.
-
-    Attributes:
-        name (str): Name of the particle species.
-        N_particles (int): Number of particles in the species.
-        charge (float): Charge of each particle.
-        mass (float): Mass of each particle.
-        weight (float): Weighting factor for the particles.
-        T (float): Temperature of the particle species.
-        v1, v2, v3 (array-like): Velocity components of the particles.
-        x1, x2, x3 (array-like): Position components of the particles.
-        dx, dy, dz (float): Spatial resolution in each dimension.
-        x_wind, y_wind, z_wind (float): Domain size in each dimension.
-        zeta1, zeta2, eta1, eta2, xi1, xi2 (float): Subcell positions for charge conservation.
-        bc (str): Boundary condition type ('periodic' or 'reflecting').
-        update_x, update_y, update_z (bool): Flags to update position in respective dimensions.
-        update_vx, update_vy, update_vz (bool): Flags to update velocity in respective dimensions.
-        update_pos (bool): Flag to update particle positions.
-        update_v (bool): Flag to update particle velocities.
-        shape (int): Shape factor for the particles (1 for first order, 2 for second order)
-
-    Methods:
-        get_name(): Returns the name of the particle species.
-        get_charge(): Returns the total charge of the particles.
-        get_number_of_particles(): Returns the number of particles in the species.
-        get_temperature(): Returns the temperature of the particle species.
-        get_velocity(): Returns the velocity components of the particles.
-        get_position(): Returns the position components of the particles.
-        get_mass(): Returns the total mass of the particles.
-        get_subcell_position(): Returns the subcell positions for charge conservation.
-        get_resolution(): Returns the spatial resolution in each dimension.
-        get_shape(): Returns the shape factor of the particles.
-        get_index(): Computes and returns the particle indices in the grid.
-        set_velocity(v1, v2, v3): Sets the velocity components of the particles.
-        set_position(x1, x2, x3): Sets the position components of the particles.
-        set_mass(mass): Sets the mass of the particles.
-        set_weight(weight): Sets the weight of the particles.
-        calc_subcell_position(): Calculates and returns the subcell positions.
-        kinetic_energy(): Computes and returns the kinetic energy of the particles.
-        momentum(): Computes and returns the momentum of the particles.
-        periodic_boundary_condition(x_wind, y_wind, z_wind): Applies periodic boundary conditions.
-        reflecting_boundary_condition(x_wind, y_wind, z_wind): Applies reflecting boundary conditions.
-        update_position(dt): Updates the positions of the particles based on their velocities and boundary conditions.
-        tree_flatten(): Flattens the object for serialization.
-        tree_unflatten(aux_data, children): Reconstructs the object from flattened data.
-    """
-
-
-
-    def __init__(self, name, N_particles, charge, mass, T, v1, v2, v3, x1, x2, x3, \
-            xwind, ywind, zwind, dx, dy, dz, weight=1, x_bc="periodic", y_bc="periodic", \
-                z_bc="periodic", update_x=True, update_y=True, update_z=True, \
-                update_vx=True, update_vy=True, update_vz=True, update_pos=True, update_v=True, shape=1, dt = 0):
-        self.name = name
-        self.N_particles = N_particles
-        self.charge = charge
-        self.mass = mass
-        self.weight = weight
-        self.T = T
-        self.v1 = v1
-        self.v2 = v2
-        self.v3 = v3
-        self.dx = dx
-        self.dy = dy
-        self.dz = dz
-        self.x_wind = xwind
-        self.y_wind = ywind
-        self.z_wind = zwind
-        self.half_x_wind = 0.5 * xwind
-        self.half_y_wind = 0.5 * ywind
-        self.half_z_wind = 0.5 * zwind
-        self.x_bc = x_bc
-        self.y_bc = y_bc
-        self.z_bc = z_bc
-        self.x_periodic = x_bc == 'periodic'
-        self.x_reflecting = x_bc == 'reflecting'
-        self.y_periodic = y_bc == 'periodic'
-        self.y_reflecting = y_bc == 'reflecting'
-        self.z_periodic = z_bc == 'periodic'
-        self.z_reflecting = z_bc == 'reflecting'
-        # boundary conditions for each dimension
-        self.update_x = update_x
-        self.update_y = update_y
-        self.update_z = update_z
-        self.update_vx = update_vx
-        self.update_vy = update_vy
-        self.update_vz = update_vz
-        self.update_pos = update_pos
-        self.update_v   = update_v
-        self.shape = shape
-        self.dt = dt
-
-        self.x1 = x1
-        self.x2 = x2
-        self.x3 = x3
-
-    def get_name(self):
-        return self.name
-
-    def get_charge(self):
-        return self.charge*self.weight
-
-    def get_number_of_particles(self):
-        return self.N_particles
-
-    def get_temperature(self):
-        return self.T
-
-    def get_velocity(self):
-        return self.v1, self.v2, self.v3
-
-    def get_forward_position(self):
-        return self.x1, self.x2, self.x3
-
-    def get_position(self):
-        x1_back = self.x1 - self.v1 * self.dt / 2
-        x2_back = self.x2 - self.v2 * self.dt / 2
-        x3_back = self.x3 - self.v3 * self.dt / 2
-        # moving the forward positions back half a time step to get x_t positions
-
-        # boundary conditions
-        if self.x_bc == 'periodic':
-            x1_back = jnp.where(x1_back > self.x_wind/2,  x1_back - self.x_wind, \
-                            jnp.where(x1_back < -self.x_wind/2, x1_back + self.x_wind, x1_back))
-        # apply boundary conditions to the x position of the particles
-
-        if self.y_bc == 'periodic':
-            x2_back = jnp.where(x2_back > self.y_wind/2,  x2_back - self.y_wind, \
-                            jnp.where(x2_back < -self.y_wind/2, x2_back + self.y_wind, x2_back))
-        # apply boundary conditions to the y position of the particles
-
-        if self.z_bc == 'periodic':
-            x3_back = jnp.where(x3_back > self.z_wind/2,  x3_back - self.z_wind, \
-                            jnp.where(x3_back < -self.z_wind/2, x3_back + self.z_wind, x3_back))
-        # apply boundary conditions to the z position of the particles
-
-        return x1_back, x2_back, x3_back
-
-    def get_mass(self):
-        return self.mass*self.weight
-
-    def get_resolution(self):
-        return self.dx, self.dy, self.dz
-
-    def get_shape(self):
-        return self.shape
-
-    def get_index(self):
-        return compute_index(self.x1, self.dx, self.x_wind), compute_index(self.x2, self.dy, self.y_wind), compute_index(self.x3, self.dz, self.z_wind)
-
-    def set_velocity(self, v1, v2, v3):
-        if self.update_v:
-            if self.update_vx:
-                self.v1 = v1
-            if self.update_vy:
-                self.v2 = v2
-            if self.update_vz:
-                self.v3 = v3
-
-    def set_position(self, x1, x2, x3):
-        self.x1 = x1
-        self.x2 = x2
-        self.x3 = x3
-
-    def set_mass(self, mass):
-        self.mass = mass
-
-    def set_weight(self, weight):
-        self.weight = weight
-
-    def kinetic_energy(self):
-        v2 = jnp.square(self.v1) + jnp.square(self.v2) + jnp.square(self.v3)
-        # compute the square of the velocity
-        return 0.5 * self.weight * self.mass *  jnp.sum(v2)
-
-    def momentum(self):
-        return self.mass * self.weight * jnp.sum(jnp.sqrt(self.v1**2 + self.v2**2 + self.v3**2))
-    
-    def boundary_conditions(self):
-        x1, x2, x3 = self.x1, self.x2, self.x3
-        v1, v2, v3 = self.v1, self.v2, self.v3
-
-        x1, v1 = apply_axis_boundary_condition(x1, v1, self.x_wind, self.half_x_wind, self.x_periodic, self.x_reflecting)
-        x2, v2 = apply_axis_boundary_condition(x2, v2, self.y_wind, self.half_y_wind, self.y_periodic, self.y_reflecting)
-        x3, v3 = apply_axis_boundary_condition(x3, v3, self.z_wind, self.half_z_wind, self.z_periodic, self.z_reflecting)
-
-        self.x1, self.x2, self.x3 = x1, x2, x3
-        self.v1, self.v2, self.v3 = v1, v2, v3
-   
-    def update_position(self):
-        if self.update_pos:
-            if self.update_x:
-                self.x1      = self.x1 + self.v1 * self.dt
-                # update the x position of the particles
-
-            if self.update_y:
-                self.x2      = self.x2 + self.v2 * self.dt
-                # update the y position of the particles
-
-            if self.update_z:
-                self.x3      = self.x3 + self.v3 * self.dt
-                # update the z position of the particles
-
-    def tree_flatten(self):
-        children = (
-            self.v1, self.v2, self.v3, \
-            self.x1, self.x2, self.x3, \
-        )
-
-        aux_data = (
-            self.name, self.N_particles, self.charge, self.mass, self.T, \
-            self.x_wind, self.y_wind, self.z_wind, self.dx, self.dy, self.dz, \
-            self.weight, self.x_bc, self.y_bc, self.z_bc, self.update_pos, self.update_v, \
-            self.update_x, self.update_y, self.update_z, self.update_vx, self.update_vy, \
-            self.update_vz, self.shape, self.dt
-        )
-        return children, aux_data
-
-    @classmethod
-    def tree_unflatten(cls, aux_data, children):
-        v1, v2, v3, x1, x2, x3 = children
-
-
-        name, N_particles, charge, mass, T, x_wind, y_wind, z_wind, dx, dy, \
-            dz, weight, x_bc, y_bc, z_bc, update_pos, update_v, update_x, update_y, update_z, \
-                update_vx, update_vy, update_vz, shape, dt  = aux_data
-
-
-        obj = cls(
-            name=name,
-            N_particles=N_particles,
-            charge=charge,
-            mass=mass,
-            T=T,
-            x1=x1,
-            x2=x2,
-            x3=x3,
-            v1=v1,
-            v2=v2,
-            v3=v3,
-            xwind=x_wind,
-            ywind=y_wind,
-            zwind=z_wind,
-            dx=dx,
-            dy=dy,
-            dz=dz,
-            weight=weight,
-            x_bc=x_bc,
-            y_bc=y_bc,
-            z_bc=z_bc,
-            update_x=update_x,
-            update_y=update_y,
-            update_z=update_z,
-            update_vx=update_vx,
-            update_vy=update_vy,
-            update_vz=update_vz,
-            update_pos=update_pos,
-            update_v=update_v,
-            shape=shape,
-            dt=dt
-        )
-
-        return obj
